+++ apps/recnet-api/src/modules/slack/slack.type.ts
@@ -1,0 +1,5 @@
+export type SendSlackResult = {
+  success: boolean;
+  skip?: boolean;
+  userId?: string;
+};