+++ apps/recnet-api/src/modules/user/entities/user.subscription.entity.ts
@@ -1,0 +1,10 @@
+import { ApiProperty } from "@nestjs/swagger";
+import { Channel, SubscriptionType } from "@prisma/client";
+
+export class Subscription {
+  @ApiProperty()
+  type: SubscriptionType;
+
+  @ApiProperty()
+  channels: Channel[];
+}